Added extra trajectory file [ci skip]
